+++ CHANGELOG.md
@@ -6,7 +6,8 @@
   * Create out of the box configuration `config/assets.config.php.dist`
   * #29 - pass more arguments to filter constructor
   * #74 - now only assets match in request are build
-  * use cache buster only when cache is enabled
+  * #80 - In debug mode, asset collection is move as separate file, not as single file.
+  * use cache buster only when cache is enabled or have cache buster enabled
 
 2013-06-10
 ----------